third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/security/mixedContent/insecure-fetch-on-dedicated-worker-allowed.https-expected.txt

CONSOLE WARNING: line 1: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/insecure-fetch-on-dedicated-worker.js' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*'. This content should also be served over HTTPS.
This test starts a dedicated worker that loads insecure data via Fetch API. We should trigger a mixed content callback but allow the load.

The test passes if the result is "LOADED".

Result: LOADED
